fix api 2

+++ src/app/api/project/delete/route.ts
@@ -8,11 +8,23 @@ export async function DELETE (request: Request){
     const body = await request.json();
 
     try {
+        if (!token) {
+            return NextResponse.json(
+                {
+                    message: "Cookie is undefined or has no value",
+                },
+                {
+                    status: 400,
+                }
+            );
+        }
+
+        const {value} = token
         const { uid } = body
         const result = axios.delete(`${process.env.NEXT_PUBLIC_API_URL}/api/projects/${uid}`, {
             headers: {
                 "Content-Type": "application/json",
-                'Authorization': `Bearer ${token.value}`
+                'Authorization': `Bearer ${value}`
             }
         })
         const { data } = await result
